+++ task01/src/by/tsvirko/task01/dao/IntegerReading.java
@@ -1,5 +1,5 @@
 package by.tsvirko.task01.dao;
 
 public interface IntegerReading {
-    int readNumber();
+    long readNumber();
 }
