+++ task01/src/by/tsvirko/task01/service/ReaderAccess.java
@@ -3,7 +3,7 @@ package by.tsvirko.task01.service;
 import java.util.Scanner;
 
 public interface ReaderAccess {
-    int fileReader(Scanner scanner, String filePath);
+    long fileReader(Scanner scanner, String filePath);
 
-    int consoleReader();
+    long consoleReader();
 }
